+++ tests/Unit/CostumeError.php
@@ -1,5 +1,6 @@
 <?php
 
+use Validator\Rule;
 use Validator\Validator;
 
 it('can change supported language', function () {
@@ -23,4 +24,40 @@ it('change supported language after run validator doesn\'t perform anything', fu
     $val->lang('id');
 
     expect($err)->toBe(['test' => 'The Test field is required']);
+});
+
+it('can create costume error', function () {
+    Rule::set_error_message('required', '{field} can\'t be null');
+
+    $val = new Validator(['tets' => null]);
+    $val->test->required();
+
+    expect($val->get_error())->test->toEqual('Test can\'t be null');
+});
+
+it('can create costume error (containt \'not\' method)', function () {
+    Rule::set_error_message('required', '{field} can\'t be null');
+
+    $val = new Validator(['test' => 'null']);
+    $val->test->not()->required();
+
+    expect($val->get_error())->test->toEqual('Test can\'t be null');
+});
+
+it('can create costume error multy', function () {
+    Rule::set_error_messages([
+        'required' => '{field} can\'t be null',
+        'min_len'  => '{field} less that 2',
+    ]);
+
+    $val = new Validator([
+        'tets' => null, 'test2' => 'abc',
+    ]);
+    $val->test->required();
+    $val->field('test2')->min_len(4);
+
+    expect($val->get_error())->toMatchArray([
+        'test'  => 'Test can\'t be null',
+        'test2' => 'Test2 less that 2',
+    ]);
 });